+++ tests/Database/DatabaseUtilTest.php
@@ -373,19 +373,19 @@ class DatabaseUtilTest extends TestCaseEnvironment
         $this->assertCount(2, $result);
         $this->assertSame('(test REGEXP (?) OR test REGEXP (?))', $result[0]);
         $this->assertCount(2, $result[1]);
-        $this->assertSame("':\"A\"'", $result[1][0]);
+        $this->assertSame(':"A"', $result[1][0]);
 
         $instance = $this->createTestInstance();
         $result = $instance->createWhereForSerializedBlob('test', ['A', 'B'], $instance::SQL_CONDITION_AND);
         $this->assertCount(2, $result);
         $this->assertSame('((test REGEXP (?)) AND (test REGEXP (?)))', $result[0]);
         $this->assertCount(2, $result[1]);
-        $this->assertSame("':\"A\"'", $result[1][0]);
+        $this->assertSame(':"A"', $result[1][0]);
 
         $result = $instance->createWhereForSerializedBlob('test', ['A', 'B'], $instance::SQL_CONDITION_OR, ['inline_values' => true]);
         $this->assertCount(2, $result);
-        $this->assertSame("(test REGEXP (':\"A\"') OR test REGEXP (':\"B\"'))", $result[0]);
+        $this->assertSame('(test REGEXP (:"A") OR test REGEXP (:"B"))', $result[0]);
         $this->assertCount(2, $result[1]);
-        $this->assertSame("':\"A\"'", $result[1][0]);
+        $this->assertSame(':"A"', $result[1][0]);
     }
 }